requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langgraph
streamlit
wikipedia
dotenv
python-dotenv
faiss-cpu
sentence-transformers
transformers
torch
langchain-text-splitters
langchain-faiss